D Latches, Rising Edge Detectors, and XOr.
Various fixes for things.

New bug discovered - due to order of operations, there needs to be a way to remove power from wires. Therefore, solution is this - every component will generate a unique hash based on its X, Y, type, and dir, along with the x and y of the output and power.  Everything powered by that output will gain that fingerprint. Equal outputs from other components will add to the array of fingerprints a wire will have. If a wire has zero fingerprints, its value should be -1. If it has several fingerprints assigned to it, its power level should be the highest of all the power levels in the fingerprints.